+++ src/Concerns/HandlesTextStreaming.php
@@ -392,7 +392,7 @@ trait HandlesTextStreaming
             }
 
             if (filled($schema)) {
-                $body['response_format'] = $this->buildResponseFormat();
+                $body['response_format'] = $this->buildResponseFormat($schema);
             }
 
             if (! is_null($options?->maxTokens)) {